Update configuration settings for PHP8+
