+++ src/FileTest.php
@@ -15,8 +15,6 @@ use League\Uri\Exceptions\SyntaxError;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @group file
- * @group uri
  * @coversDefaultClass \League\Uri\Uri
  */
 final class FileTest extends TestCase
