--- src/ResolverTest.php
@@ -1,86 +1,0 @@
-<?php
-
-/**
- * League.Uri (https://uri.thephpleague.com)
- *
- * (c) Ignace Nyamagana Butera <nyamsprod@gmail.com>
- *
- * For the full copyright and license information, please view the LICENSE
- * file that was distributed with this source code.
- */
-
-namespace League\Uri;
-
-use PHPUnit\Framework\TestCase;
-
-/**
- * @group modifer
- * @coversDefaultClass \League\Uri\UriResolver
- */
-final class ResolverTest extends TestCase
-{
-    public function testResolveLetThrowResolvedInvalidUri(): void
-    {
-        $http = Uri::createFromString('http://example.com/path/to/file');
-        $ftp = Http::createFromString('ftp://a/b/c/d;p');
-        $res = UriResolver::resolve($ftp, $http);
-        self::assertEquals($res, $ftp);
-    }
-
-    /**
-     * @dataProvider resolveProvider
-     */
-    public function testCreateResolve(string $base_uri, string $uri, string $expected): void
-    {
-        self::assertSame($expected, (string) UriResolver::resolve(
-            Uri::createFromString($uri),
-            Http::createFromString($base_uri)
-        ));
-    }
-
-    public function resolveProvider(): array
-    {
-        $base_uri = 'http://a/b/c/d;p?q';
-
-        return [
-            'base uri'                => [$base_uri, '',              $base_uri],
-            'scheme'                  => [$base_uri, 'http://d/e/f',  'http://d/e/f'],
-            'path 1'                  => [$base_uri, 'g',             'http://a/b/c/g'],
-            'path 2'                  => [$base_uri, './g',           'http://a/b/c/g'],
-            'path 3'                  => [$base_uri, 'g/',            'http://a/b/c/g/'],
-            'path 4'                  => [$base_uri, '/g',            'http://a/g'],
-            'authority'               => [$base_uri, '//g',           'http://g'],
-            'query'                   => [$base_uri, '?y',            'http://a/b/c/d;p?y'],
-            'path + query'            => [$base_uri, 'g?y',           'http://a/b/c/g?y'],
-            'fragment'                => [$base_uri, '#s',            'http://a/b/c/d;p?q#s'],
-            'path + fragment'         => [$base_uri, 'g#s',           'http://a/b/c/g#s'],
-            'path + query + fragment' => [$base_uri, 'g?y#s',         'http://a/b/c/g?y#s'],
-            'single dot 1'            => [$base_uri, '.',             'http://a/b/c/'],
-            'single dot 2'            => [$base_uri, './',            'http://a/b/c/'],
-            'single dot 3'            => [$base_uri, './g/.',         'http://a/b/c/g/'],
-            'single dot 4'            => [$base_uri, 'g/./h',         'http://a/b/c/g/h'],
-            'double dot 1'            => [$base_uri, '..',            'http://a/b/'],
-            'double dot 2'            => [$base_uri, '../',           'http://a/b/'],
-            'double dot 3'            => [$base_uri, '../g',          'http://a/b/g'],
-            'double dot 4'            => [$base_uri, '../..',         'http://a/'],
-            'double dot 5'            => [$base_uri, '../../',        'http://a/'],
-            'double dot 6'            => [$base_uri, '../../g',       'http://a/g'],
-            'double dot 7'            => [$base_uri, '../../../g',    'http://a/g'],
-            'double dot 8'            => [$base_uri, '../../../../g', 'http://a/g'],
-            'double dot 9'            => [$base_uri, 'g/../h' ,       'http://a/b/c/h'],
-            'mulitple slashes'        => [$base_uri, 'foo////g',      'http://a/b/c/foo////g'],
-            'complex path 1'          => [$base_uri, ';x',            'http://a/b/c/;x'],
-            'complex path 2'          => [$base_uri, 'g;x',           'http://a/b/c/g;x'],
-            'complex path 3'          => [$base_uri, 'g;x?y#s',       'http://a/b/c/g;x?y#s'],
-            'complex path 4'          => [$base_uri, 'g;x=1/./y',     'http://a/b/c/g;x=1/y'],
-            'complex path 5'          => [$base_uri, 'g;x=1/../y',    'http://a/b/c/y'],
-            'dot segments presence 1' => [$base_uri, '/./g',          'http://a/g'],
-            'dot segments presence 2' => [$base_uri, '/../g',         'http://a/g'],
-            'dot segments presence 3' => [$base_uri, 'g.',            'http://a/b/c/g.'],
-            'dot segments presence 4' => [$base_uri, '.g',            'http://a/b/c/.g'],
-            'dot segments presence 5' => [$base_uri, 'g..',           'http://a/b/c/g..'],
-            'dot segments presence 6' => [$base_uri, '..g',           'http://a/b/c/..g'],
-            'origin uri without path' => ['http://h:b@a', 'b/../y',   'http://h:b@a/y'],
-        ];
-    }
-}
