+++ src/WsTest.php
@@ -15,8 +15,6 @@ use League\Uri\Exceptions\SyntaxError;
 use PHPUnit\Framework\TestCase;
 
 /**
- * @group ws
- * @group uri
  * @coversDefaultClass \League\Uri\Uri
  */
 class WsTest extends TestCase
